+++ cli/tests/test_operations.rs
@@ -1778,6 +1778,37 @@ fn test_op_diff_sibling() {
     - qpvuntsm hidden b1ca67e2 (empty) B
     [EOF]
     ");
+
+    // no graph
+    let output = work_dir.run_jj([
+        "op",
+        "diff",
+        "--at-op",
+        p2_op_id,
+        "--from",
+        p2_op_id,
+        "--to",
+        p1_op_id,
+        "--summary",
+        "--no-graph",
+    ]);
+    insta::assert_snapshot!(output, @r"
+    From operation: 252ff3a5a0e6 (2001-02-03 08:05:12) describe commit e8849ae12c709f2321908879bc724fdb2ab8a781
+      To operation: 7bba3a63b73b (2001-02-03 08:05:11) new empty commit
+
+    Changed commits:
+    - qpvuntsm hidden b1ca67e2 (empty) B
+    + mzvwutvl 08c63613 (empty) A
+    + zsuskuln 47b9525e A.2
+    A file2
+    + kkmpptxz 6c70a4f7 A.1
+    A file1
+
+    Changed working copy default@:
+    + mzvwutvl 08c63613 (empty) A
+    - qpvuntsm hidden b1ca67e2 (empty) B
+    [EOF]
+    ");
 }
 
 #[test]