+++ CHANGELOG.md
@@ -29,7 +29,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Templates now support a `.split(separator, [limit])` method on strings to
   split a string into a list of substrings.
 
-* `jj log -G` is now available as a short form of `jj log --no-graph`.
+* `-G` is now available as a short form of `--no-graph` in `jj log`, `jj evolog`,
+  `jj op log`, `jj op show` and `jj op diff`.
 
 * `jj metaedit` now accepts `-m`/`--message` option to non-interactively update
   the change description.